factor out to text
